+++ src/user/top_artists.rs
@@ -1,59 +1,53 @@
-use crate::error::{Error, LastFMError};
-use crate::user::User;
-use crate::{Client, RequestBuilder};
 use serde::Deserialize;
 use std::marker::PhantomData;
 
+use crate::{
+    error::{Error, LastFMError},
+    model::{Attributes, Image},
+    user::User,
+    Client, RequestBuilder,
+};
+
 /// The main top artists structure.
 ///
-/// This is splitted off into two areas: One, the attributes (used
-/// for displaying various user-associated attributes), and two,
-/// the user's top artists.
+/// This is splitted off into two areas: One, the attributes (used for displaying various
+/// user-associated attributes), and two, the user's top artists.
 ///
-/// For details on the attributes available, refer to [Attributes]. For
-/// details on the artist information available, refer to [Artist].
+/// For details on the attributes available, refer to [Attributes]. For details on the artist information
+/// available, refer to [Artist].
 #[derive(Debug, Deserialize)]
 pub struct TopArtists {
+    /// A [Vec] array containing a user's Top Artists.
     #[serde(rename = "artist")]
     pub artists: Vec<Artist>,
+    /// Various internal API attributes.
     #[serde(rename = "@attr")]
     pub attrs: Attributes,
 }
 
 #[derive(Debug, Deserialize)]
-pub struct Attributes {
-    pub page: String,
-    pub total: String,
-    pub user: String,
-    #[serde(rename = "perPage")]
-    pub per_page: String,
-    #[serde(rename = "totalPages")]
-    pub total_pages: String,
-}
-
-#[derive(Debug, Deserialize)]
 pub struct Artist {
+    /// Attributes associated with the artist.
     #[serde(rename = "@attr")]
     pub attrs: ArtistAttributes,
+    /// The MusicBrainz ID for the artist.
     pub mbid: String,
-    pub playcount: String,
+    /// How many times the user has scrobbled the artist.
+    #[serde(rename = "playcount")]
+    pub scrobbles: String,
+    /// The name of the artist.
     pub name: String,
+    /// The Last.fm URL for the artist.
     pub url: String,
+    /// The main images of the artist.
     #[serde(rename = "image")]
     pub images: Vec<Image>,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct ArtistAttributes {
+    /// Where the artist is ranked in the user's profile.
     pub rank: String,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct Image {
-    #[serde(rename = "size")]
-    pub image_size: String,
-    #[serde(rename = "#text")]
-    pub image_url: String,
 }
 
 impl TopArtists {
@@ -63,22 +57,29 @@ impl TopArtists {
     }
 }
 
-/// Allows users to specify the period of which they'd like to retrieve
-/// top artist data for.
+/// Allows users to specify the period of which they'd like to retrieve top artist data for.
 pub enum Period {
     Overall,
+    /// Retrieves data collected over the past 7 days.
     SevenDays,
+    /// Retrieves data collected over the past month.
     OneMonth,
+    /// Retrieves data collected over the past three months.
     ThreeMonths,
+    /// Retrieves data collected over the past six months.
     SixMonths,
+    /// Retrieves data collected over the past twelve months.
     TwelveMonths,
+    /// Retrieves data collected over the past year. Retrieves the same data as [TwelveMonths]. This
+    /// only serves as a shorter shortcut to the same thing.
+    ///
+    /// [TwelveMonths]: Period::TwelveMonths
     OneYear,
 }
 
 impl ToString for Period {
-    /// Converts the given period to a string. In most cases,
-    /// you won't have to use this yourself. Period durations
-    /// will usually be automatically converted to their string
+    /// Converts the given period to a string. In most cases, you won't have to use this
+    /// yourself. Period durations will usually be automatically converted to their string
     /// form when fed to the `with_period` parameter function.
     fn to_string(&self) -> String {
         match self {
@@ -87,11 +88,7 @@ impl ToString for Period {
             Self::OneMonth => String::from("1month"),
             Self::ThreeMonths => String::from("3month"),
             Self::SixMonths => String::from("6month"),
-            // TwelveMonths and OneYear are exactly the same,
-            // it just allows users to choose which one they'd
-            // like to use, depending on verbosity / simplicity.
-            Self::TwelveMonths => String::from("12month"),
-            Self::OneYear => String::from("12month"),
+            Self::TwelveMonths | Self::OneYear => String::from("12month"),
         }
     }
 }
@@ -119,7 +116,5 @@ impl<'a> RequestBuilder<'a, TopArtists> {
 }
 
 impl<'a> Client {
-    pub async fn top_artists(&'a mut self, user: &str) -> RequestBuilder<'a, TopArtists> {
-        TopArtists::build(self, user).await
-    }
+    pub async fn top_artists(&'a mut self, user: &str) -> RequestBuilder<'a, TopArtists> { TopArtists::build(self, user).await }
 }